requirements.txt
numpy>=1.24,<2.0
scipy>=1.10
wfdb>=4.1
tensorflow>=2.13,<2.17
pyyaml>=6.0
matplotlib>=3.7
scikit-learn>=1.3
